admin: split out partition routes
