+++ src/v/redpanda/CMakeLists.txt
@@ -48,6 +48,7 @@ v_cc_library(
     admin/security.cc
     admin/recovery.cc
     admin/transaction.cc
+    admin/partition.cc
     admin/usage.cc
     admin/kafka.cc
     admin/util.cc
